tear down LCAO_hamilt.h and .cpp, DONE.
